+++ esphome/components/mcp3204/__init__.py
@@ -2,14 +2,10 @@ import esphome.codegen as cg
 import esphome.config_validation as cv
 from esphome.components import spi
 from esphome.const import CONF_ID
-from esphome.core import CORE
 
 DEPENDENCIES = ["spi"]
-AUTO_LOAD = ["sensor"]
 MULTI_CONF = True
 CODEOWNERS = ["@rsumner"]
-
-CONF_MCP3204 = "mcp3204"
 
 mcp3204_ns = cg.esphome_ns.namespace("mcp3204")
 MCP3204 = mcp3204_ns.class_("MCP3204", cg.Component, spi.SPIDevice)
@@ -25,6 +21,3 @@ async def to_code(config):
     var = cg.new_Pvariable(config[CONF_ID])
     await cg.register_component(var, config)
     await spi.register_spi_device(var, config)
-
-    if CORE.is_esp32:
-        cg.add_library("SPI", None)
