+++ esphome/components/mcp3204/sensor/mcp3204_sensor.h
@@ -1,0 +1,32 @@
+#pragma once
+
+#include "esphome/components/sensor/sensor.h"
+#include "esphome/components/voltage_sampler/voltage_sampler.h"
+#include "esphome/core/component.h"
+#include "esphome/core/hal.h"
+
+#include "../mcp3204.h"
+
+namespace esphome {
+namespace mcp3204 {
+
+class MCP3204Sensor : public PollingComponent,
+                      public Parented<MCP3204>,
+                      public sensor::Sensor,
+                      public voltage_sampler::VoltageSampler {
+ public:
+  MCP3204Sensor(uint8_t pin, float reference_voltage);
+
+  void set_reference_voltage(float reference_voltage) { reference_voltage_ = reference_voltage; }
+  void update() override;
+  void dump_config() override;
+  float get_setup_priority() const override;
+  float sample() override;
+
+ protected:
+  uint8_t pin_;
+  float reference_voltage_;
+};
+
+}  // namespace mcp3204
+}  // namespace esphome
